requirements.txt
requests>=2.31.0
Pillow>=10.0.0
reportlab>=4.0
python-barcode>=0.15.1
streamlit>=1.28.0
numpy>=1.24.0
click>=8.1.0
pytest>=7.0
python-dotenv>=1.0.0
